refactor(domain): Separate Entity and Repository in Domain
