+++ app/Like/Domain/Entities/Like.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Like\Domain\Entities;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use App\Review\Domain\Entities\Review;
+use App\Profile\Domains\Entities\Users;
+
+class Like extends Model
+{
+  use HasFactory;
+
+  protected $table = 'likes';
+  protected $fillable = ['review_id', 'user_id', 'created_at', 'updated_at'];
+
+  public function review()
+  {
+    return $this->belongsTo(Review::class);
+  }
+
+  public function user()
+  {
+    return $this->belongsTo(Users::class);
+  }
+}
